+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.10.6] - 2022-05-22
+### Added
+- Block mining speed traits (`silentgear:block_mining_speed`) which increase the mining speed for blocks in given tags
+- Greedy trait, which increases the mining speed on ores (`forge:ores`)
+- Replaced Soft with Greedy on Blaze Gold main material
+
 ## [2.10.5] - 2022-05-13
 ### Fixed
 - Mattocks not tilling dirt [#502]